Split intersection_preview.rs into intersection_preview/ subdirectory
Split crates/rendering/src/intersection_preview.rs (525 lines) into focused modules
under intersection_preview/ for better maintainability.
